+++ CHANGELOG.md
@@ -6,6 +6,9 @@ to [Semantic Versioning][].
 This document follows the conventions laid out in [Keep a CHANGELOG][].
 
 ## [Unreleased][]
+
+## [Release][2.0.5] [2.0.5][2.0.4-2.0.5] - 2023-03-16
+
 ### Added
 - New Keywords for click and hold implemented
     - Click Hold
@@ -384,7 +387,11 @@ General bugfixing from keywords and improvements from artifact building and test
 
 [github.io]: https://gdatasoftwareag.github.io/robotframework-flaui
 
-[unreleased]: https://github.com/GDATASoftwareAG/robotframework-flaui/compare/2.0.4...main
+[unreleased]: https://github.com/GDATASoftwareAG/robotframework-flaui/compare/2.0.5...main
+
+[2.0.5]: https://github.com/GDATASoftwareAG/robotframework-flaui/releases/tag/2.0.5
+
+[2.0.4-2.0.5]: https://github.com/GDATASoftwareAG/robotframework-flaui/compare/2.0.4...2.0.5
 
 [2.0.4]: https://github.com/GDATASoftwareAG/robotframework-flaui/releases/tag/2.0.4